requirements.txt
streamlit==1.31.1
groq==0.4.2
pytesseract==0.3.10
PyPDF2==3.0.1
sympy==1.12
torch==2.2.1
sentence-transformers==2.4.0
faiss-cpu==1.7.4
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3